Setup basic Book scaffold
To prepare for testing we need to have _something_ to actually test.
This sets up a basic scaffold for a model we're calling `Book`. To keep
things simple it has a `title` and `author`. As the focus of this test
project is system tests we're keeping things simple by not setting up
validations.

We likely only need a single view, but as we learn more about
stubbing/recording system tests we may want to expand functionality. To
keep future commits focus on the actual specs / setup we're leaving all
of the generated controller / views in place.