Setup System specs
Setting up basic system tests is fairly straight forward. We need to
include the `capybara` and `selenium0webdriver` gems. We include them in
the development and test groups because that is what Rails will do by
default if you included system tests from the start; which we didn't
because we are using RSpec.

This modifies the `Gemfile` slightly by not auto requiring these gems.
This is to show how I would set this up on a real project. This is
especially true for a project which doesn't use spring. I wouldn't want
these gems to get loaded unless they are needed; to keep isolated
testing of other types (e.g. model specs) fast. We setup a hook so that
when RSpec sees our first system test it loads the gems and performs
basic setup.

As I have chrome already configured on this system we use the selenium
chrome driver; instead of just the default `selenium` driver. At the
time of writing this I have Chrome Version 64.0.3282.167. You will want
at least a recent version of chrome as a quick online search will show
that some older versions had a few issues with system tests (though that
is mostly for the headless driver which we'll get to later).