Setup some client and server remote requests
This sets up some basic client side and server side remote requests. On
the client side we use JSONP, through jQuery, to hit the public
non-authenticated GitHub API. This returns the meta information for the
API. As it's non-authenticated it makes a good test case for simplistic
JSONP usage. On the server side, we just use the built-in `Net::HTTP`
lib as it's convenient.

Note that these base system tests **will** hit the network. In order to
run them you must be connected. We make this clear by tagging them with
conditional RSpec metadata which only runs if a networks is available.