Add puffing-billy for browser proxy caching
Webmock and VCR are great for handling network request stubs for app
code. However, system tests use an external web browser and requests
made the browser will not hit either of those tools (e.g. JSONP requests
in javascript, redirects to other domains, etc.). In order to
stub/record those requests we need a different tool. Enter
puffing-billy, a tool which sets up a proxy to intercept the requests
from the browser.

For the most part this works great. However, as are illustrated by these
specs there is some weirdness around in spec requests through `visit`
and how the ports work. To send things to the browser the `visit` call
for system tests always appends a port for HTTP/HTTPS requests if one is
not set. This breaks stubbing of requests with puffing-billy which
checks the port; though it doesn't break the caches because they ignore
the ports.

Additionally, when writing the specs in isolation everything ran fine.
But once all specs were run there were random failures. I traced this
back to the chrome browser cache intercepting various redirects /
`visit` requests and serving them from it's local cache. To make the
specs behave as intended we setup a custom driver which disables the
chrome cache. For now we only use this for those specs which use
puffing-billy cache (which changes between specs) and not stubs (which
explicitly set responses and did not seem to have issues).

One thing I have noticed is that chrome tries to get the favicon for
most sites. I haven't found a way to tell puffing-billy to just return
those as no data without recording everything. This is a bit annoying
when recording new specs as you have to pick through the various cached
recordings to focus just on the ones you need.