Integrate webmock with puffing-billy
Overall this was relatively painless. Due to the threading design,
webmock is available on the puffing-billy proxy instance. This means we
can use it for all of general request stubs instead of the puffing-billy
stub system if we choose. However, this got a little weird when running
multiple drivers (some which we want to have webmock with and others we
don't). In a normal app I would probably just want webmock everywhere
and this wouldn't be an issue, but given we are trying to test behavior
of various combinations of items we need to add some hooks to ensure
it's on or off early before things get spun up.

Webmock sits behind the puffing-billy cache system, so if desired it can
be temporarily disabled to collect remote requests to cache, then
re-enabled to handle things that you don't want to cach (like favicons);
alternatively we can just configure webmock to allow the desired remote
requests through so they can get cached.

Also, since webmock is loaded on the proxy instance it can handle all of
the cases which puffing-billy cannot; such as server side requests. But
given the awkward way you have to stub JSONP it may be better to use
puffing-billy for those and leave everything else to webmock.

One downside is that the errors thrown when webmock stops an unknown
request are not well handled by the puffing-billy proxy. This causes a
stack dump to be displayed in the spec output, outside of the actual
error reporting. And occasionally it will cause Ruby to segfault.