Integrate VCR with puffing-billy (mostly)
This was a bit painful to get setup. The pain stemmed from two issues:

  - https://github.com/oesmith/puffing-billy/pull/228
  - https://github.com/vcr/vcr/issues/615

The first issue is that the current release of puffing-billy doesn't
handle errors raised by the proxy requests. This causes the errors to
bubble up crashing the proxy process. This then causes the chromedriver
to hang forever. Until this is fixed we hot patch the code. While we
only faced this when integrating VCR, as it raises errors by default
when things aren't stubbed, it's possible webmock or any other library
can cause a similar issue. So we apply the patch to any system spec.

The 2nd issue is specific to VCR and how it interacts with Event Machine
used by puffing-billy. Unfortunately there doesn't seem to be away
around this at the moment. The webmock hook to drop the zip header is a
fairly heavy :hammer: and there's no guarantee that this will work with
binary or other VCR cassettes - I didn't test that. Hopefully, the issue
will be fixed at some point or that this hammer is enough.

In order to support the proxy, which is on our localhost, we have to
tell VCR to allow this. Also, this repo modifies the actual recorded
responses, so we tell VCR to update the content length header to match
our custom modifications.

This is **mostly** working as everything but JSONP works. If your JSONP
isn't dynamic this is probably all you will need. However, this is using
jQuery for the client side JSONP requests. The way we are using jQuery
causes a new dynamic method name. This breaks VCR in two ways:

  - the URI won't match across requests (param values are different)
  - we can't modify the response to use the dynamic param

Until we get that sorted out the spec is marked as pending.