Implement JSONP for VCR
This was a tough one. Inspiration for this solution came from:

  - https://github.com/vcr/vcr/issues/284
  - https://github.com/nbibler/vcr-284

The issue is that VCR's hooks don't give you an easy way to get both the
real request URI and the response together at any of the times you can
make modifications.

Per the above issue, the real URI isn't available for the custom
cassettes placeholder hook; but only for the stub. This can be confusing
because with the real request the cassette isn't recorded yet and when
the block is called to parse which text to replace it appears in the
URI. But then once the cassette is used the URI for the request is that
in the recorded cassette. Which in the case of JSONP will either be the
original request and or the placeholder text itself depending on you
configure the replacements.

The other VCR hooks aren't very helpful. Specifically for this case, on
playback the order of the hooks, with argument types, are:

  - `before_playback(VCR::HTTPInteraction::HookAware, VCR::Cassette)`
  - `before_http_request(VCR::Request::Typed)`
  - `after_http_request(VCR::Request::Typed, VCR::Response)`

The `before_playback` does have the `HTTPInteraction`, but as stated
above that is just the cassette interaction and doesn't have the real
URI. The `before_http_request` does provide the real URI requested, but
you don't have access to the interaction or response. Based on the above
definitions it would seem then that the `after_http_request` is what we
want, and ideally it would be, but with the webmock hook you can't
modify the response.

Well you can, but it's ignored. It's hard to trace what's going on and
when but it seems like the `after_http_request` is given a copy of the
response. Well it looks like at least the [body is
duplicated](https://github.com/vcr/vcr/blob/v4.0.0/lib/vcr/structs.rb#L68-L73).

At this point our options seem to be:

  - monkey patch webmock to either:
    - provide the real URI to `before_playback`
    - provide the real response to `before_http_request`
    - provide the real response to `after_http_request`
  - write a custom library hook / request handler
  - find another way to patch the response

Personally, I think the best long term solution is to update
`before_playback` to also provide the real URI. However, there are some
issues with this because `before_playback` is just called once for the
spec; well once for the cassette. This is then used for all of the
requests.  So this change would require a large change to how VCR works
internally.

Next is the `before_http_request`. This is called prior to matching the
request to any of the cassette recordings. So at this point we don't
have a matching response. Based on the [code comments](https://github.com/vcr/vcr/blob/v4.0.0/lib/vcr/request_handler.rb#L14-L21)
this is intentional to allow the hook to modify the request or even
change the cassette used.

So really the only "public" API we have is the `after_http_request`
hook. But as noted above it doesn't allow us to make a change to the
response which is then used afterwards. Modifying the `VCR::Response`
struct seems like it will be messy given the age of the original comment
issue. And adding a new hook isn't an option either. It looks like it
could be possible to monkey patch (instead of writing a whole new
handler) how the `stubbed_response` is handled. As the handler has
access to both the request and response it could perform the
substitutions. But I'd like to try to stay way from all that if
possible without VCR maintainer support.

The only other thing that seems related are these webmock hooks. That's
what VCR is using internally for the `after_http_request` callback. It
seems that at least with our combination of VCR and webmock this works.
And it seems like it's using "public" APIs so that's a plus too.